feat: rename cloud to connect
