--- packages/cloud/src/components/index.ts
@@ -1,2 +1,0 @@
-export * from "./cloud";
-export * from "./withCloud";
