--- packages/cloud/src/contexts/cloud/ICloudContext.ts
@@ -1,7 +1,0 @@
-export type ICloudContext = {
-    baseUrl?: string;
-    clientId?: string;
-    resourcesName?: string;
-};
-
-export type CloudConfigProvider = Required<ICloudContext>;
