--- packages/cloud/src/interfaces/index.ts
@@ -1,1 +1,0 @@
-export * from "../contexts/cloud/ICloudContext";
