+++ packages/connect/.npmignore
