+++ packages/connect/jest.config.js
