+++ packages/connect/package.json
@@ -1,0 +1,60 @@
+{
+  "version": "1.0.0",
+  "license": "MIT",
+  "main": "dist/index.js",
+  "typings": "dist/index.d.ts",
+  "private": false,
+  "files": [
+    "dist",
+    "src"
+  ],
+  "engines": {
+    "node": ">=10"
+  },
+  "scripts": {
+    "start": "tsup --watch --format esm,cjs,iife --legacy-output",
+    "build": "tsup --format esm,cjs,iife --minify --legacy-output",
+    "test": "jest --passWithNoTests --runInBand",
+    "prepare": "npm run build"
+  },
+  "name": "@refinedev/connect",
+  "description": "refine connect",
+  "author": "refine",
+  "module": "dist/esm/index.js",
+  "peerDependencies": {
+    "@refinedev/core": "^4.0.0",
+    "@refinedev/sdk": "^1.0.0",
+    "react": "^17.0.0 || ^18.0.0",
+    "react-dom": "^17.0.0 || ^18.0.0",
+    "@types/react": "^17.0.0 || ^18.0.0",
+    "@types/react-dom": "^17.0.0 || ^18.0.0"
+  },
+  "devDependencies": {
+    "@refinedev/core": "^4.1.0",
+    "@refinedev/sdk": "^1.1.0",
+    "@esbuild-plugins/node-resolve": "^0.1.4",
+    "@types/lodash": "^4.14.171",
+    "@tanstack/react-query": "^4.10.1",
+    "@types/jest": "^29.2.4",
+    "jest": "^29.3.1",
+    "jest-environment-jsdom": "^29.3.1",
+    "react-router-dom": "^6.8.1",
+    "ts-jest": "^29.0.3",
+    "tslib": "^2.3.1",
+    "tsup": "^5.11.13"
+  },
+  "dependencies": {
+    "lodash-es": "^4.17.21",
+    "lodash": "^4.17.21",
+    "react-query": "^3.19.2"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/refinedev/refine.git",
+    "directory": "packages/react-hook-form"
+  },
+  "gitHead": "829f5a516f98c06f666d6be3e6e6099c75c07719",
+  "publishConfig": {
+    "access": "public"
+  }
+}
