+++ packages/connect/src/components/index.ts
@@ -1,0 +1,2 @@
+export * from "./connect";
+export * from "./withConnect";
