+++ packages/connect/src/contexts/connect/IConnectContext.ts
@@ -1,0 +1,7 @@
+export type IConnectContext = {
+    baseUrl?: string;
+    clientId?: string;
+    resourcesName?: string;
+};
+
+export type ConnectConfigProvider = Required<IConnectContext>;
