+++ packages/connect/src/contexts/connect/index.tsx
@@ -1,0 +1,22 @@
+import React from "react";
+
+import { IConnectContext } from "./IConnectContext";
+
+export const ConnectContext = React.createContext<IConnectContext>({});
+
+export const ConnectContextProvider: React.FC<
+    IConnectContext & {
+        children: React.ReactNode;
+    }
+> = ({ baseUrl, clientId, children }) => {
+    return (
+        <ConnectContext.Provider
+            value={{
+                baseUrl,
+                clientId,
+            }}
+        >
+            {children}
+        </ConnectContext.Provider>
+    );
+};
