+++ packages/connect/src/hooks/index.ts
@@ -1,0 +1,6 @@
+export * from "./useSdk";
+export * from "./useAuthProviderWithConnectConfig";
+export * from "./useAuthConfig";
+export * from "./useResourcesConfig";
+export * from "./useConnectQuery";
+export * from "./useConnectMutation";
