+++ packages/connect/src/hooks/useAuthConfig/index.ts
