+++ packages/connect/src/hooks/useResourcesConfig/index.ts
