+++ packages/connect/tsconfig.declarations.json
